fix supporting delete many for 3rd party

+++ packages/keyv/src/index.js
@@ -154,6 +154,15 @@ class Keyv extends EventEmitter {
 		let keyPrefixed = this._getKeyPrefix(key);
 		if (Array.isArray(key)) {
 			keyPrefixed = this._getKeyPrefixArray(key);
+			if (store.deleteMany === 'function'){
+				const promises = [];
+				for (const key of keys) {
+					promises.push(this.delete(key));
+				}
+
+				return Promise.allSettled(promises)
+					.then(values => values.every(x => x.value === true));
+			}
 			return Promise.resolve()
 				.then(() => store.deleteMany(key));
 		}
